Correct signedness handling of DispatchTimeInterval
Replicate e32016ffd765020c696bc5d1e42606074305f2db
from the Darwin overlay to the wrapping overlay.

+++ src/swift/IO.swift
@@ -88,7 +88,7 @@ public extension DispatchIO {
 	}
 
 	public func setInterval(interval: DispatchTimeInterval, flags: IntervalFlags = []) {
-		dispatch_io_set_interval(self.__wrapped, interval.rawValue, flags.rawValue)
+		dispatch_io_set_interval(self.__wrapped, UInt64(interval.rawValue), flags.rawValue)
 	}
 
 	public func close(flags: CloseFlags = []) {
